+++ Pages/InfoPage.xaml.cs
@@ -1,4 +1,6 @@
-﻿namespace EnlightenMAUI
+﻿using EnlightenMAUI.Models;
+
+namespace EnlightenMAUI
 {
     public partial class InfoPage : ContentPage
     {
@@ -6,6 +8,12 @@
         {
             InitializeComponent();
         }
-    }
 
+        private async void Button_Clicked(object sender, EventArgs e)
+        {
+            Settings settings = Settings.getInstance();
+            await Browser.OpenAsync(settings.companyURL);
+
+        }
+    }
 }