+++ Pages/ClipboardPage.xaml.cs
@@ -1,0 +1,10 @@
+﻿namespace EnlightenMAUI;
+
+[XamlCompilation(XamlCompilationOptions.Compile)]
+public partial class ClipboardPage : ContentPage
+{
+    public ClipboardPage()
+    {
+        InitializeComponent();
+    }
+}