+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - 2024-04-15 0.0.4
     - added enlighten_logo
     - About -> Info
+    - InfoViewModel -> code-behind
 - 2024-04-14 0.0.3
     - Telerik chart loading
     - imported from EnlightenMobile